refactor(station-detail): extract sections from station_detail_screen (Refs #563)
Split the 431-LOC station_detail_screen into four cohesive widgets so the
screen file stays under the 300-LOC presentation-layer cap:

- StationDetailAppBarActions — directions/alert/scan-QR/report/favorite
  cluster plus the create-alert + scan-payment dispatch logic.
- StationBrandHeader — brand logo + name + "Independent station"
  subtitle (#482) block.
- StationPricesSection — "Prices" heading, per-fuel tiles, and the
  LogFillUpButton CTA.
- station_brand_helpers.dart — shared hasRealBrand/isIndependentSentinel
  predicates so the screen title and the brand header agree.

Pure extraction — no behaviour change, no ARB touch, no provider surface
change. Screen drops from 431 → 149 LOC.
